+++ frontend/src/modules/article/editor/articleEditorSaga.js
@@ -1,7 +1,7 @@
 import { takeLatest } from 'redux-saga/effects';
-import createRequestSaga from '../../services/createRequestSaga';
+import createRequestSaga from '../../../services/createRequestSaga';
 import { WRITE_ARTICLE } from './articleEditorTypes';
-import * as articleAPI from '../../services/api/articleAPI';
+import * as articleAPI from '../../../services/api/articleAPI';
 
 const writeArticleSaga = createRequestSaga(WRITE_ARTICLE, articleAPI.write);
 export function* articleEditorSaga() {
